feat: Pipes
Introduce a "pipe" concept for processes and other entities that provide
a input/output streams.

Makes it possible to write things like this:

    exec("dmesg") | exec("grep net") | exec("sort")

Also, combine script functions as intermediate steps in the pipeline.

Introduces just enough async stuff to make it work with minimal impact
on the rest of the code base.